backend/requirements.txt
# Core
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pyjwt==2.8.0

# Database
psycopg2-binary==2.9.7
pgvector==0.2.0

# AI / NLP (LOCK CỨNG – KHÔNG ĐƯỢC ĐỔI)
sentence-transformers==2.2.2
transformers==4.26.1
huggingface-hub==0.16.4
torch==1.13.1
numpy==1.24.3
